#153 Move get_dtype and get_name to core.specs module
